Follow the established pattern for command registration
